swal alert en delete favs

+++ client/src/components/Favs/Favs.jsx
@@ -2,6 +2,7 @@ import React, { useEffect } from "react";
 import { BsArrowLeft } from "react-icons/bs";
 import { useDispatch, useSelector } from "react-redux";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 import { deleteFav } from "../../actions";
 import style from "./style.module.css";
 
@@ -11,14 +12,32 @@ export default function Favs(){
 
     const handleDelFavs = (e) =>{
         e.preventDefault()
-        dispatch(deleteFav(e.target.id))
+        Swal.fire({
+            title: '¿Estás seguro?',
+            text: "Esta acción no se revertirá!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Si! eliminar!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+              Swal.fire(
+                  'Eliminado!',
+                  'Su Pokemon ha sido eliminado de favoritos.',
+                  'success',
+                  dispatch(deleteFav(e.target.id))
+              )
+            }
+          })
+        
     }
         
 return(
     <div>
 
      <span className={style.span} >
-                <h2 className={style.h2} >My pokemons</h2>
+                <h2 className={style.h2} >Pokemons favoritos</h2>
                 <Link to={"/home"} className={style.link} ><BsArrowLeft/></Link>
      </span>
      <div className={style.conteiner} >
